simplify frontend react initialization
